Created a reducer fir the pressed button in a different file
